enhancement: move thor & disco to runtime package
